+++ src/lib.rs
@@ -3,6 +3,6 @@ pub mod models;
 mod whatsapp_client;
 
 pub use crate::whatsapp_client::WhatsappClient;
-pub use error::WhatsappError;
+pub use error::{ApiError, ApiErrorData, WhatsappError};
 
 pub const WHATSAPP: &str = "whatsapp";